requirements.txt
Django==1.4
wsgiref==0.1.2
django-tastypie
